fixed the issue with this not submitting
It was getting confused with having two inputs with the same name and
id.
Solved issue by having it remove the old button row, captcha image and
input box.
Also was having difficulty with a string that had escaped special
characters, changed it to a single quote string, that fixed it.